refactor: merge json init file to handle
